+++ test/api/subscription.test.ts
@@ -19,16 +19,19 @@ suite('SubscriptionItem tests', async () => {
     test('Tree should show a subscription item for each Azure subscription', async () => {
         const { mockResources, subscriptionItems } = await mockResourcesAndGetSubscriptionItems();
         mockResources.subscriptions.forEach((subscription) => {
-            assert.ok(subscriptionItems?.find((node) => node.id === subscription.id));
+            assert.ok(
+                subscriptionItems?.find((node) => node.subscription.subscriptionId === subscription.subscriptionId),
+                `Subscription ${subscription.id} not found in the tree. Found: ${subscriptionItems.map((node) => node.subscription.subscriptionId).join(', ')}`
+            );
         });
 
         assert.strictEqual(subscriptionItems.length, mockResources.subscriptions.length, `There should be ${mockResources.subscriptions.length} subscription nodes. Found ${subscriptionItems.length}.`);
     });
 
-    test('SubscriptionItem.id should be in the "/subscriptions/<subscription id>" format', async () => {
+    test('SubscriptionItem.id should be in the "/accounts/<account id>/tenants/<tenant id>/subscriptions/<subscription id>" format', async () => {
         const { subscriptionItems } = await mockResourcesAndGetSubscriptionItems();
         subscriptionItems.forEach((subscriptionItem) => {
-            assert.match(subscriptionItem.id, /^\/subscriptions\/([0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12})/i);
+            assert.match(subscriptionItem.id, /^\/accounts\/[^/]+\/tenants\/[^/]+\/subscriptions\/([0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12})/i);
         });
     });